Extract FCGI logic
